fix(query)!: Add field name validation to QueryAction
BREAKING CHANGE:  application code that supplied unrecognized field names to query actions in the past used to be up to the specific backend module what to do with them (e.g., RDBMS would fail, but memory would silently ignore).  Now, before we get into the backend module, we will validate field names in criteria and order by elements of the filter.  This feels like the right thing to always do, and good to be standardized - but it is a (breaking) change in behavior.